update to authenticate within the tile instead of in the header
