+++ src/Components/Header/Header.js
@@ -1,6 +1,5 @@
 import React, { useEffect } from 'react';
 import styled, { keyframes } from 'styled-components';
-import base64url from 'base64url';
 import { Link as BaseLink, useLocation } from 'react-router-dom';
 import { useWalletService, WINDOW_MESSAGES } from '@provenanceio/wallet-lib';
 import { useWallet } from 'redux/hooks';
@@ -111,12 +110,7 @@ const Navigation = () => {
     setWalletLogin,
     setWalletLogout,
     walletUrl,
-    isKYC,
-    isKYCChecked,
-    getWalletKYC,
-    jwtToken,
     setJwtToken,
-    publicKeyB64,
   } = useWallet();
 
   // Get URL Query Parameters
@@ -174,27 +168,6 @@ const Navigation = () => {
     walletService.updateState();
   };
 
-  const handleKYCSign = () => {
-    const expires = Math.floor(Date.now() / 1000) + 9000; // 900s (15min)
-    const header = JSON.stringify({alg: 'ES256', typ: 'JWT'});
-    const headerEncoded = base64url(header);
-    // const addressBase64 = base64url(address);
-    const payload = JSON.stringify({sub: `${publicKeyB64},${address}`, iss: 'provenance.io', iat: expires, exp: expires});
-    const payloadEncoded = base64url(payload);
-    const jwtEncoded = base64url(`${headerEncoded}.${payloadEncoded}`);
-    // Open the wallet and sign the payload
-    walletService.sign({payload: jwtEncoded});
-    // Create window event listener (once wallet finishes)
-    walletService.addEventListener(WINDOW_MESSAGES.SIGNATURE_COMPLETE, ({ message = {} }) => {
-      const { signedPayload } = message;
-      const fullJWT = `${headerEncoded}.${payloadEncoded}.${signedPayload}`;
-      // Save token in store
-      setJwtToken(fullJWT);
-      // Use the response to send to the wallet
-      getWalletKYC({address, publicKeyB64, fullJWT});
-    }, { once: true });
-  };
-
   return (
     <HeaderWrapper>
       <LogoLink to='/' title="Provenance Applications">
@@ -202,9 +175,6 @@ const Navigation = () => {
         <LogoText>Apps</LogoText>
         {process.env.REACT_APP_ENV !== 'production' && <TestnetMessage>TESTNET</TestnetMessage>}
       </LogoLink>
-      {!isKYC && !isKYCChecked && address && !jwtToken &&
-        <Button onClick={handleKYCSign}>Check Account KYC</Button>
-      }
       {isLoggedIn ?
         <LoginStatus>
           <LoginMsg>
